Implemented a parsing algorithm for the monthly rapla response
